Fix file open deadlock on macos by delaying mutex aquisition until needed
Prior to this change, when pressing "Load HPO" on a mac, the system would
show the dialog box but with the "spinning beach ball" and the app would
be unresponsive. I think the issue may be that the file picker may need
to interact with the main thread/event loop, so having the mutex locked
ahead of time is problematic.

This just moves the mutex to directly where it is needed (when Ok from the
file picker, then to report status), and holding the mutex for a limited
time.